+++ example/test/mocktail/mocker_mocktail_test.fallback.dart
@@ -13,15 +13,24 @@ import '../models/model_b.dart' as _i4;
 
 void registerFallbackValuesAutoDetected() {
   _i1.registerFallbackValue(_$MockModel3());
+  _i1.registerFallbackValue(_$MockModel7());
   _i1.registerFallbackValue(_$MockModel());
   _i1.registerFallbackValue(_$Mock$Model());
   _i1.registerFallbackValue(_$MockModel6());
+  _i1.registerFallbackValue(_$MockObject());
+  _i1.registerFallbackValue(_$MockInvocation());
 }
 
 class _$MockModel3 extends _i1.Mock implements _i2.Model3 {}
+
+class _$MockModel7 extends _i1.Mock implements _i2.Model7 {}
 
 class _$MockModel extends _i1.Mock implements _i3.Model {}
 
 class _$Mock$Model extends _i1.Mock implements _i4.Model {}
 
 class _$MockModel6 extends _i1.Mock implements _i2.Model6 {}
+
+class _$MockObject extends _i1.Mock implements Object {}
+
+class _$MockInvocation extends _i1.Mock implements Invocation {}